backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pandas>=2.2.2
google-adk
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0